+++ CHANGELOG.md
@@ -4,6 +4,16 @@ This project adheres to [Semantic Versioning](http://semver.org/)
 
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
+## 4.0.0
+
+### Added
+ - Tracks expand with wider window size
+
+### Changed
+ - Resolution increased 2x for tracks
+
+### Fixed
+
 ## 3.1.0
 
 ### Added